llvm: fix compilation of llvm
The CMakeList was broken: shared files (q_shared) were flagged as "llvm"
files and later the linker didn't know what to do with that .bc
Reintroduced _ds_'s foo/./bar hack for that.
Updated vm_llvm  to work with my LLVM 3.2 though the opt passes
are not used by the code for now.